Replace size_t with uint64_t where necessary to avoid implicit casts between the two; Replace __mode_t with mode_t.

+++ components/core/src/GlobalSQLiteMetadataDB.cpp
@@ -235,7 +235,10 @@ void GlobalSQLiteMetadataDB::close () {
     m_is_open = false;
 }
 
-void GlobalSQLiteMetadataDB::add_archive (const string& id, size_t uncompressed_size, size_t size, const string& creator_id, size_t creation_num) {
+void GlobalSQLiteMetadataDB::add_archive (const string& id, uint64_t uncompressed_size,
+                                          uint64_t size, const string& creator_id,
+                                          uint64_t creation_num)
+{
     if (false == m_is_open) {
         throw OperationFailed(ErrorCode_NotInit, __FILENAME__, __LINE__);
     }
@@ -249,7 +252,9 @@ void GlobalSQLiteMetadataDB::add_archive (const string& id, size_t uncompressed_
     m_insert_archive_statement->reset();
 }
 
-void GlobalSQLiteMetadataDB::update_archive_size (const string& archive_id, size_t uncompressed_size, size_t size) {
+void GlobalSQLiteMetadataDB::update_archive_size (const string& archive_id,
+                                                  uint64_t uncompressed_size, uint64_t size)
+{
     if (false == m_is_open) {
         throw OperationFailed(ErrorCode_NotInit, __FILENAME__, __LINE__);
     }
